pyproject.toml
[tool.poetry]
name = "tapir"
version = "0.1.0"
description = ""
authors = ["Leon Handreke <leonh@ndreke.de>, Théophile Madet <theo.madet@posteo.net>"]

[tool.poetry.dependencies]
python = "^3.13"
Django = "^5.1.3"
django-weasyprint = "^2.3.0"
django-extensions = "^3.2.3"
django-bootstrap-datepicker-plus = "^5.0.5"
psycopg2 = "^2.9.10"
django-tables2 = "^2.7.0"
django-filter = "^23.5"
tablib = "^3.7.0"
django-phonenumber-field = "^8.0.0"
phonenumbers = "^8.13.50"
weasyprint = "^62.3"
django-select2 = "^8.2.1"
django-silk = "^5.3.1"
gunicorn = "^22.0.0"
django-environ = "^0.11.2"
django-bootstrap5 = "^23.4"
celery = "^5.4.0"
redis = "^5.2.0"
ipython = "^8.29.0"
django-chartjs = "^2.3.0"
unidecode = "^1.3.8"
python-barcode = "^0.15.1"
fabric = "^3.2.2"
requests = "^2.32.3"
django-auth-ldap = "^4.8.0"
djangorestframework = "^3.15.2"
django-vite = "^3.0.5"
django-cors-headers = "^4.6.0"
parameterized = "^0.9.0"
distinctipy = {extras = ["extras"], version = "^1.3.4"}

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
ipython = "^8.29.0"
pre-commit = "^3.8.0"
werkzeug = "^3.1.3"
pytest = "^8.3.3"
pytest-cov = "^5.0.0"
pytest-django = "^4.9.0"
pytest-sugar = "^1.0.0"
selenium = "^4.27.0"
factory-boy = "^3.3.1"
icecream = "^2.1.3"
djlint = "^1.36.1"
pypdf = "^4.3.1"
django-upgrade = "^1.22.1"
drf-spectacular = "^0.27.2"

[build-system]
requires = ["poetry-core>=1.2.0"]
build-backend = "poetry.core.masonry.api"

[tool.djlint]
profile = "django"
max_line_length = 120
ignore = "T002,H006,H021,H025,H030,H031"
use_gitignore = true


[tool.coverage.run]
omit = [
    "*/migrations/*",
     "*/tests/*",
    "*/urls.py",
    "migrate.py"
]
relative_files = true

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "tapir.settings"
python_files = ["tests.py", "test_*.py", "tests_*.py"]
testpaths = "tapir"
